% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nomads-package.R
\docType{package}
\name{nomads-package}
\alias{nomads}
\alias{nomads-package}
\title{nomads: Nomadic Wader Movement Data}
\description{
Provides satellite tracking data from nomadic pectoral sandpipers.
}
\author{
\strong{Maintainer}: Mihai Valcu \email{mihai.valcu@bi.mpg.de}

Authors:
\itemize{
  \item Mihai Valcu \email{mihai.valcu@bi.mpg.de}
}

Other contributors:
\itemize{
  \item Bart Kempenaers \email{bart.kempenaers@bi.mpg.de} [contributor]
}

}
\keyword{internal}
